Source code

Functions

Types

Base.rand — Function

rand(mpp::MarkedPoissonProcess, n::Int=1)

Generate a couple of waiting time and mark realizations from the process mpp.

Base.rand — Function

rand(pp::PoissonProcess, n::Int=1)

Generate n waiting time realizations from the process pp.

Distributions.rate — Method

rate(mpp::MarkedPoissonProcess)

Return the ground rate of the Marked Poisson Process mpp.

Distributions.rate — Method

rate(pp::PoissonProcess)

Return the rate of the Poisson Process pp.

MarkedPoissonProcesses.compose — Method

compose(pps::Vector{PoissonProcess})

Construct the composed Poisson Process from the sub-processes pps.

MarkedPoissonProcesses.condprocess — Method

condprocess(mpp::MarkedPoissonProcess; threshold::Real)

Compute the marked Poisson Process conditional that the mark is larger than "threshold".

MarkedPoissonProcesses.ground — Method

ground(mpp::MarkedPoissonProcess)

Return the ground process.

MarkedPoissonProcesses.nevents — Method

nevents(mpp::MarkedPoissonProcess, t::Real)

Compute the expected number of events in the interval (0,t) of the Marked Poisson Process mpp.

MarkedPoissonProcesses.nevents — Method

nevents(pp::PoissonProcess, t::Real)

Compute the expected number of events in the interval (0,t) of the Poisson Process pp.

MarkedPoissonProcesses.prob — Method

prob(pp::PoissonProcess, t::Real)

Compute the probability that the next event will occur before t.

MarkedPoissonProcesses.thin — Method

thin(pp::PoissonProcess, α::Real)

Compute the Poisson process pp thinned with the factor (0 < α < 1).

MarkedPoissonProcesses.weibullfit — Method

weibullfit(y::AbstractArray{<:Real})

Compute the maximum likelihood estimates of the Weibull distribution parameters with maximum likelihood

Statistics.mean — Method

mean(tpd::Truncated{Weibull{Float64}, Continuous, Float64})

Compute the mean of the truncated Weibull distribution by numerical integration.

Statistics.mean — Method

mean(mpp::MarkedPoissonProcess)

Compute the expected waiting time before the next event and the expected mark of the Marked Poisson Process mpp.

Statistics.mean — Method

mean(pp::PoissonProcess)

Compute the expected waiting time before the next event of the Poisson Process pp.

Statistics.quantile — Method

quantile(pp::PoissonProcess, p::Real)

Computation of the time $s$ for which the probability of the next event occurring in the interval $(0, s)$ is p.